refactor: refactor controller to include ResponseEntity on my routes to customize the requests responses for the user

+++ src/main/java/dev/vitorbomfim/EmployersTaskAssigner/Employer/EmployerController.java
@@ -1,5 +1,9 @@
 package dev.vitorbomfim.EmployersTaskAssigner.Employer;
 
+import org.apache.coyote.Response;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.HttpStatusCode;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -19,36 +23,62 @@ public class EmployerController {
 // Add employer (CREATE)
 
     @PostMapping("/post")
-    public EmployerDTO addEmployer(@RequestBody EmployerDTO employerDTO){
-        return employerService.addEmployer(employerDTO);
+    public ResponseEntity<String> addEmployer(@RequestBody EmployerDTO employerDTO){
+      EmployerDTO employerDT = employerService.addEmployer(employerDTO);
+        return ResponseEntity.status(HttpStatus.CREATED)
+                .body("The employer: " + employerDT.getName() + " has beed added with sucess, with ID: " + employerDT.getId());
     }
 
     // Get all employers (READ)
 
     @GetMapping("/get")
-    public List<EmployerDTO> getEmployer(){
-        return employerService.getEmployers();
+    public ResponseEntity<List<EmployerDTO>> getEmployer(){
+        List<EmployerDTO> getEmployer = employerService.getEmployers();
+        return ResponseEntity.ok(getEmployer);
     }
 
     // Search employer by id (READ)
 
     @GetMapping("/get/{id}")
-    public EmployerDTO getEmployerById(@PathVariable Long id){
-            return employerService.listEmployerById(id);
+    public ResponseEntity<?> getEmployerById(@PathVariable Long id){
+            EmployerDTO employerDTO = employerService.listEmployerById(id);
+
+            if (employerDTO != null) {
+                return ResponseEntity.ok(employerDTO);
+            } else {
+                return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                        .body("ERROR! The Employer with ID: " + id + " has not found.");
+            }
     }
 
     // Update employer data (UPDATE)
 
     @PutMapping("/update/{id}")
-    public EmployerDTO updateEmployer(@PathVariable Long id, @RequestBody EmployerDTO employerDTO){
-        return employerService.updateEmployer(id, employerDTO);
+    public ResponseEntity<?> updateEmployer(@PathVariable Long id, @RequestBody EmployerDTO employerUpdated){
+        EmployerDTO employerDTO = employerService.updateEmployer(id, employerUpdated);
+
+        if (employerDTO != null){
+            return ResponseEntity.ok(employerDTO);
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body("ERROR! The Employer with ID: " + id + " has not been found.");
+        }
     }
 
     // Delete employer (DELETE)
 
     @DeleteMapping("/delete/{id}")
-    public void deleteEmployerById(@PathVariable Long id){
-        employerService.deleteEmployerById(id);
+    public ResponseEntity<String> deleteEmployerById(@PathVariable Long id){
+
+        if (employerService.listEmployerById(id) != null){
+            employerService.deleteEmployerById(id);
+            return ResponseEntity.ok()
+                    .body("The Employer with ID: " + id + " deleted with success" );
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body("ERROR! The Employer with ID: " + id + " has not been found");
+        }
+
     }
 
 }
